requirements.txt
appnope==0.1.2
backcall==0.2.0
cffi==1.15.0
cloudpickle==1.6.0
cycler==0.11.0
Cython==0.29.24
decorator==5.1.0
fonttools==4.28.1
glfw==2.4.0
gym==0.19.0
imageio==2.10.4
importlib-resources==5.4.0
ipython==7.29.0
jedi==0.18.1
kiwisolver==1.3.2
lockfile==0.12.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
mdptoolbox-hiive==4.0.3.1
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.58
packaging==21.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycparser==2.21
pydot==1.4.2
pyglet==1.5.21
Pygments==2.10.0
pymdptoolbox==4.0b3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
traitlets==5.1.1
wcwidth==0.2.5
zipp==3.6.0
